requirements.txt
absl-py==2.1.0
asttokens==2.4.1
backcall==0.2.0
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
decorator==5.1.1
executing==2.0.1
fonttools==4.53.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.30.0
google-auth-oauthlib==1.0.0
grpcio==1.64.1
idna==3.7
importlib_metadata==7.2.1
importlib_resources==6.4.0
ipython==8.12.3
jedi==0.19.1
kiwisolver==1.4.5
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
numpy==1.24.4
oauthlib==3.2.2
opencv-python==4.10.0.84
packaging==24.1
pandas==2.0.3
parso==0.8.4
pickleshare==0.7.5
pillow==9.5.0
prompt_toolkit==3.0.47
protobuf==5.27.1
psutil==6.0.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scipy==1.10.1
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
Werkzeug==3.0.3
zipp==3.19.2